changed dartcontroller to be a compostion with Menus
changed Menus to be a compostion with storage.
